# dist/public/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "Gender": {
                "enum": [
                    "male",
                    "female",
                    "others"
                ],
                "type": "string"
            },
            "IUpdateProfileInput": {
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "country": {
                        "properties": {
                            "currency": {
                                "type": "string"
                            },
                            "callingCode": {
                                "type": "string"
                            },
                            "code": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "currency",
                            "callingCode",
                            "code",
                            "name"
                        ],
                        "type": "object"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string"
                    },
                    "gender": {
                        "$ref": "#/components/schemas/Gender"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "IUpdatePasswordInput": {
                "properties": {
                    "oldPassword": {
                        "type": "string"
                    },
                    "newPassword": {
                        "type": "string"
                    }
                },
                "required": [
                    "oldPassword",
                    "newPassword"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "IUpdatePinInput": {
                "properties": {
                    "oldPin": {
                        "type": "string"
                    },
                    "newPin": {
                        "type": "string"
                    }
                },
                "required": [
                    "oldPin",
                    "newPin"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
        "/user": {
            "get": {
                "operationId": "GetUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": []
            },
            "patch": {
                "operationId": "UpdateUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IUpdateProfileInput"
                            }
                        }
                    }
                }
            }
        },
        "/user/password": {
            "patch": {
                "operationId": "UpdatePassword",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IUpdatePasswordInput"
                            }
                        }
                    }
                }
            }
        },
        "/user/pin": {
            "patch": {
                "operationId": "UpdatePin",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IUpdatePinInput"
                            }
                        }
                    }
                }
            }
        },
        "/user/enable-notifications": {
            "post": {
                "operationId": "EnableNotifications",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IEnableNotificationsInput"
                            }
                        }
                    }
                }
            }
        },
        "/user/all-banks/{country}": {
            "get": {
                "operationId": "GetAllBanks",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "country",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/user/rewards": {
            "get": {
                "operationId": "GetRewards",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "items": {
                                                "$ref": "#/components/schemas/Reward"
                                            },
                                            "type": "array"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": []
            }
        },
...
    },
    "info": {
        "license": {
            "name": "ISC"
        },
        "contact": {},
        "title": "MD Exchange API",
        "description": "API for managing MD Exchange operations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "localhost dev"
        }
    ]
}